Avoid implicit cast from size_t to int

+++ cgltf_write.h
@@ -144,7 +144,7 @@ static void cgltf_write_line(cgltf_write_context* context, const char* line)
 	}
 	cgltf_write_indent(context);
 	CGLTF_SPRINTF("%s", line);
-	int last = strlen(line) - 1;
+	size_t last = strlen(line) - 1;
 	if (line[0] == ']' || line[0] == '}')
 	{
 		context->needs_comma = 1;
